requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai==1.10.0
aiohttp==3.9.1
httpx==0.26.0
PyGithub==2.1.1
atlassian-python-api==3.41.0
jira==3.5.2
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
redis==5.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
aiohttp
alembic
atlassian-python-api
azure-identity
azure-keyvault-secrets
botbuilder-core
botbuilder-schema
fastapi
httpx
jira
openai
opentelemetry-api
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
prometheus-client
psycopg2-binary
pydantic
pydantic-settings
PyGithub
python-dotenv
python-multipart
redis
sqlalchemy
uvicorn
